# gin/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ticker/chart/{ticker}": {
            "get": {
                "description": "2000 성공\n2001 empty response\n4001 missing count\n4002 date format error yyyy-mm-dd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "종목 차트 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "마지막 날짜",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ticker.OneTickerChartResponse"
                            }
                        }
                    }
                }
            }
        },
...
        "/ticker/search": {
            "get": {
                "description": "2000 성공\n4001 offset and count are int type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "종목 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "검색할 종목명 , 코드",
                        "name": "word",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가지고 있는 list.length",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ticker.SearchTickerResponse"
                            }
                        }
                    }
                }
            }
        },
        "/ticker/{ticker}": {
            "get": {
                "description": "2000 성공\n4001 required count\n4002 cannot found list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "종목 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticker.OneTickerResponse"
                        }
                    }
                }
            }
        },
        "/user/nickname": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "2000 성공\n2001 nickname is exists\n4001 required nickname\n4101 required login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "닉네임 변경",
                "parameters": [
                    {
                        "description": "nickname",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.NicknameBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "auth.UserResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "created_at": {
                    "description": "가입일",
                    "type": "string"
                },
                "email": {
                    "description": "email 마우스 호버시에만 보이게",
                    "type": "string"
                },
                "nickname": {
                    "description": "nickname",
                    "type": "string"
                },
                "profile_image": {
                    "description": "google login시 받아오는 profile image",
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "ticker.OneTickerChartResponse": {
            "type": "object",
            "properties": {
                "close": {
                    "description": "종가(현재가)",
                    "type": "number"
                },
                "date": {
                    "description": "날짜",
                    "type": "string"
                },
                "high": {
                    "description": "고가",
                    "type": "number"
                },
                "low": {
                    "description": "저가",
                    "type": "number"
                },
                "open": {
                    "description": "시가",
                    "type": "number"
                },
                "percent": {
                    "description": "등락률",
                    "type": "string"
                },
                "volume": {
                    "description": "거래량",
                    "type": "string"
                }
            }
        },
        "ticker.OneTickerResponse": {
            "type": "object",
            "properties": {
                "bps": {
                    "type": "number"
                },
                "chart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ticker.OneTickerChartResponse"
                    }
                },
                "div": {
                    "type": "number"
                },
                "dps": {
                    "type": "number"
                },
                "eps": {
                    "type": "number"
                },
                "homepage": {
                    "description": "홈페이지 (로고 url)",
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "listing_date": {
                    "description": "상장일",
                    "type": "string"
                },
                "market": {
                    "description": "0 : 코스피 , 1 : 코스닥",
                    "type": "integer"
                },
                "market_cap": {
                    "description": "시가총액",
                    "type": "number"
                },
                "name": {
                    "description": "종목명",
                    "type": "string"
                },
                "pbr": {
                    "description": "표시",
                    "type": "number"
                },
                "per": {
                    "description": "표시",
                    "type": "number"
                },
                "representative": {
                    "description": "대표자 (표시할 필요없습니다.)",
                    "type": "string"
                },
                "sector": {
                    "description": "표시",
                    "type": "string"
                },
                "settle_month": {
                    "description": "결산월",
                    "type": "string"
                },
                "symbol": {
                    "description": "index",
                    "type": "string"
                }
            }
        },
        "ticker.SearchTickerResponse": {
            "type": "object",
            "properties": {
                "homepage": {
                    "description": "홈페이지 (로고 url)",
                    "type": "string"
                },
                "market": {
                    "description": "0 코스피 , 1 코스닥",
                    "type": "integer"
                },
                "name": {
                    "description": "종목명",
                    "type": "string"
                },
                "symbol": {
                    "description": "코드",
                    "type": "string"
                }
            }
        },
        "user.NicknameBody": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}